split InventoryCommands into Server and Client parts
