+++ src/main/java/io/singularitynet/MissionHandlers/InventoryCommandsImplementation.java
@@ -47,7 +47,7 @@ import org.apache.logging.log4j.Logger;
  */
 public class InventoryCommandsImplementation extends CommandGroup implements IMalmoMessageListener
 {
-    private static final Logger LOGGER = LogManager.getLogger();
+    private static final Logger LOGGER = LogManager.getLogger(InventoryCommandsImplementation.class.getName());
 
     @Override
     public void onMessage(MalmoMessageType messageType, Map<String, String> data) {
@@ -56,213 +56,9 @@ public class InventoryCommandsImplementation extends CommandGroup implements IMa
 
     @Override
     public void onMessage(MalmoMessageType messageType, Map<String, String> data, ServerPlayerEntity player) {
-        InventoryMessage msg = new InventoryMessage(data);
-        runCommand(msg, player);
+        throw new RuntimeException("Calling server-side message handler on client " + messageType.toString());
     }
 
-    public static class InventoryMessage extends MalmoMessage {
-        String invA;
-        String invB;
-        int slotA;
-        int slotB;
-        boolean combine;
-        BlockPos containerPos;
-
-        public InventoryMessage(Map<String, String> data) {
-            super(MalmoMessageType.CLIENT_INVENTORY_CHANGE, "inventory");
-            this.getData().putAll(data);
-            this.invA = data.get("invA");
-            this.invB = data.get("invB");
-            this.slotA = Integer.valueOf(data.get("slotA"));
-            this.slotB = Integer.valueOf(data.get("slotB"));
-            this.combine = Boolean.valueOf(data.get("combine"));
-            this.containerPos = null;
-            if (data.containsKey("containerPos"))
-                this.containerPos = BlockPos.fromLong(Long.valueOf(data.get("containerPos")));
-        }
-
-        public InventoryMessage(List<Object> params, boolean combine) {
-            super(MalmoMessageType.CLIENT_INVENTORY_CHANGE, "inventory");
-            this.invA = (String) params.get(0);
-            this.slotA = (Integer) params.get(1);
-            this.invB = (String) params.get(2);
-            this.slotB = (Integer) params.get(3);
-            this.combine = combine;
-            if (params.size() == 5) {
-                this.containerPos = (BlockPos) params.get(4);
-                this.getData().put("containerPos", String.valueOf(this.containerPos.asLong()));
-            }
-            this.getData().put("invA", this.invA);
-            this.getData().put("slotA", String.valueOf(this.slotA));
-            this.getData().put("slotB", String.valueOf(this.slotB));
-            this.getData().put("invB", this.invB);
-            this.getData().put("combine", String.valueOf(combine));
-        }
-    }
-
-    public static class InventoryChangeMessage {
-        public ItemStack itemsGained = null;
-        public ItemStack itemsLost = null;
-
-        public InventoryChangeMessage(ItemStack itemsGained, ItemStack itemsLost) {
-            this.itemsGained = itemsGained;
-            this.itemsLost = itemsLost;
-        }
-    }
-
-    static ItemStack[] swapSlots(ServerPlayerEntity player, String lhsInv, int lhs, String rhsInv, int rhs, BlockPos containerPos)
-    {
-        PlayerInventory container = null;
-        String containerName = "";
-
-        PlayerInventory lhsInventory = lhsInv.equals("inventory") ? player.getInventory() : (lhsInv.equals(containerName) ? container : null);
-        PlayerInventory rhsInventory = rhsInv.equals("inventory") ? player.getInventory() : (rhsInv.equals(containerName) ? container : null);
-        if (lhsInventory == null || rhsInventory == null)
-            return null; // Source or dest container not available.
-        if (rhs < 0 || lhs < 0)
-            return null; // Out of bounds.
-        if (lhs >= lhsInventory.size() || rhs >= rhsInventory.size())
-            return null; // Out of bounds.
-
-        ItemStack srcStack = lhsInventory.removeStack(lhs);
-        ItemStack dstStack = rhsInventory.removeStack(rhs);
-        LOGGER.info(String.format("setting %d to %s", lhs, dstStack.toString()));
-        LOGGER.info(String.format("setting %d to %s", rhs, srcStack.toString()));
-        lhsInventory.insertStack(lhs, dstStack);
-        rhsInventory.insertStack(rhs, srcStack);
-        lhsInventory.updateItems();
-        rhsInventory.updateItems();
-        if (lhsInventory.player instanceof ServerPlayerEntity) {
-            ServerPlayerEntity entity = ((ServerPlayerEntity)(lhsInventory.player));
-            entity.networkHandler.sendPacket(new ScreenHandlerSlotUpdateS2CPacket(-2, 0, lhs, lhsInventory.getStack(lhs)));
-            LOGGER.info(String.format("have now %d to %s", lhs, lhsInventory.getStack(lhs).toString()));
-        }
-        if (rhsInventory.player instanceof ServerPlayerEntity) {
-            ServerPlayerEntity entity = ((ServerPlayerEntity)(rhsInventory.player));
-            entity.networkHandler.sendPacket(new ScreenHandlerSlotUpdateS2CPacket(-2, 0, rhs, rhsInventory.getStack(rhs)));
-            LOGGER.info(String.format("have now %d to %s", rhs, rhsInventory.getStack(rhs).toString()));
-        }
-        if (lhsInventory != rhsInventory)
-        {
-            // Items have moved between our inventory and the foreign inventory - may need to trigger
-            // rewards for collecting / discarding.
-            ItemStack[] returnStacks = new ItemStack[2];
-            ItemStack stackBeingLost = (lhsInventory == player.getInventory()) ? srcStack : dstStack;
-            ItemStack stackBeingGained = (lhsInventory == player.getInventory()) ? dstStack : srcStack;
-            if (stackBeingGained != null)
-                returnStacks[0] = stackBeingGained.copy();
-            if (stackBeingLost != null)
-                returnStacks[1] = stackBeingLost.copy();
-            return returnStacks;
-        }
-        return null;
-    }
-
-    private static InventoryChangeMessage runCommand(InventoryMessage message, ServerPlayerEntity player){
-            ItemStack[] changes = null;
-            if (message.combine) {
-                LOGGER.info("combine");
-                changes = combineSlots(player, message.invA, message.slotA, message.invB, message.slotB, message.containerPos);
-            }
-            else {
-                LOGGER.info("swapSlots");
-                changes = swapSlots(player, message.invA, message.slotA, message.invB, message.slotB, message.containerPos);
-            }
-            if (changes != null)
-                return new InventoryChangeMessage(changes[0], changes[1]);
-            return null;
-    }
-
-    static ItemStack[] combineSlots(ServerPlayerEntity player, String invDst, int dst, String invAdd, int add, BlockPos containerPos)
-    {
-        Inventory container = null;
-        String containerName = "";
-        /*
-        if (containerPos != null)
-        {
-            TileEntity te = player.world.getTileEntity(containerPos);
-            if (te != null && te instanceof TileEntityLockableLoot)
-            {
-                containerName = ObservationFromFullInventoryImplementation.getInventoryName((IInventory)te);
-                container = (IInventory)te;
-            }
-            else if (te != null && te instanceof TileEntityEnderChest)
-            {
-                containerName = ObservationFromFullInventoryImplementation.getInventoryName(player.getInventoryEnderChest());
-                container = player.getInventoryEnderChest();
-            }
-        }*/
-        Inventory dstInv = invDst.equals("inventory") ? player.getInventory() : (invDst.equals(containerName) ? container : null);
-        Inventory addInv = invAdd.equals("inventory") ? player.getInventory() : (invAdd.equals(containerName) ? container : null);
-        if (dstInv == null || addInv == null)
-            return null; // Source or dest container not available.
-
-        ItemStack dstStack = dstInv.getStack(dst);
-        ItemStack addStack = addInv.getStack(add);
-
-        if (addStack == null)
-            return null; // Combination is a no-op.
-
-        ItemStack[] returnStacks = null;
-
-        if (dstStack == null) // Do a straight move - nothing to combine with.
-        {
-            if (dstInv != addInv)
-            {
-                // Items are moving between our inventory and the foreign inventory - may need to trigger
-                // rewards for collecting / discarding.
-                returnStacks = new ItemStack[2];
-                ItemStack stackBeingLost = (addInv == player.getInventory()) ? addStack : null;
-                ItemStack stackBeingGained = (dstInv == player.getInventory()) ? addStack : null;
-                if (stackBeingGained != null)
-                    returnStacks[0] = stackBeingGained.copy();
-                if (stackBeingLost != null)
-                    returnStacks[1] = stackBeingLost.copy();
-            }
-            dstInv.setStack(dst, addStack);
-            addInv.setStack(add, null);
-            return returnStacks;
-        }
-
-        // Check we can combine. This logic comes from InventoryPlayer.storeItemStack():
-        boolean itemsMatch = dstStack.getItem() == addStack.getItem();
-        boolean dstCanStack = dstStack.isStackable() && dstStack.getCount() < dstStack.getMaxCount() &&
-                dstStack.getCount() < dstInv.getMaxCountPerStack();
-        boolean subTypesMatch = true; //!dstStack.getHasSubtypes() || dstStack.getMetadata() == addStack.getMetadata();
-        boolean tagsMatch = ItemStack.areEqual(dstStack, addStack);
-        if (itemsMatch && dstCanStack && subTypesMatch && tagsMatch)
-        {
-            // We can combine, so figure out how much we have room for:
-            int limit = Math.min(dstStack.getMaxCount(), dstInv.getMaxCountPerStack());
-            int room = limit - dstStack.getCount();
-            ItemStack itemsTransferred = dstStack.copy();
-            if (addStack.getCount() > room)
-            {
-                // Not room for all of it, so shift across as much as possible.
-                addStack.decrement(room);
-                dstStack.increment(room);
-                itemsTransferred.setCount(room);
-            }
-            else
-            {
-                // Room for the whole lot, so empty out the add slot.
-                dstStack.increment(addStack.getCount());
-                itemsTransferred.setCount(addStack.getCount());
-                addInv.removeStack(add);//setInventorySlotContents(add, null);
-            }
-            if (dstInv != addInv)
-            {
-                // Items are moving between our inventory and the foreign inventory - may need to trigger
-                // rewards for collecting / discarding.
-                returnStacks = new ItemStack[2];
-                if (dstInv == player.getInventory())
-                    returnStacks[0] = itemsTransferred; // We're gaining them
-                else
-                    returnStacks[1] = itemsTransferred; // We're losing them
-            }
-        }
-        return returnStacks;
-    }
 
     @Override
     protected boolean onExecute(String verb, String parameter, MissionInit missionInit)
@@ -270,6 +66,7 @@ public class InventoryCommandsImplementation extends CommandGroup implements IMa
 
         if (verb.equalsIgnoreCase(InventoryCommand.SWAP_INVENTORY_ITEMS.value()))
         {
+            LOGGER.debug("SWAP_INVENTORY_ITEMS: " + parameter);
             if (parameter != null && parameter.length() != 0)
             {
                 List<Object> params = new ArrayList<Object>();
@@ -277,7 +74,8 @@ public class InventoryCommandsImplementation extends CommandGroup implements IMa
                 {
                     // All okay, so create a swap message for the server:
                     // InventoryChangeMessage msg = runCommand(new InventoryMessage(params, false));
-                    ClientPlayNetworking.send(NetworkConstants.CLIENT2SERVER, (new InventoryMessage(params, false)).toBytes());
+                    LOGGER.debug("Sending SWAP_INVENTORY_ITEMS message to server");
+                    ClientPlayNetworking.send(NetworkConstants.CLIENT2SERVER, (new InventoryCommandsImplementationServer.InventoryMessage(params, false)).toBytes());
                     return true;
                 }
                 else
@@ -293,7 +91,8 @@ public class InventoryCommandsImplementation extends CommandGroup implements IMa
                 {
                     // All okay, so create a combine message for the server:
                     // runCommand(new InventoryMessage(params, false));
-                    ClientPlayNetworking.send(NetworkConstants.CLIENT2SERVER, (new InventoryMessage(params, false)).toBytes());
+                    LOGGER.debug("Sending COMBINE_INVENTORY_ITEMS message to server");
+                    ClientPlayNetworking.send(NetworkConstants.CLIENT2SERVER, (new InventoryCommandsImplementationServer.InventoryMessage(params, false)).toBytes());
                     return true;
                 }
                 else
@@ -302,6 +101,7 @@ public class InventoryCommandsImplementation extends CommandGroup implements IMa
         }
         else if (verb.equalsIgnoreCase(InventoryCommand.DISCARD_CURRENT_ITEM.value()))
         {
+            LOGGER.debug("discarding currently selected item");
             MinecraftClient.getInstance().player.dropSelectedItem(false);
             // This we can do on the client side:
             // Minecraft.getMinecraft().player.dropItem(false);  // false means just drop one item - true means drop everything in the current stack.
@@ -415,17 +215,6 @@ public class InventoryCommandsImplementation extends CommandGroup implements IMa
         InventoryCommands iparams = (InventoryCommands) params;
         setUpAllowAndDenyLists(iparams.getModifierList());
         return true;
-    }
-
-    @Override
-    public void install(MissionInit missionInit)
-    {
-        SidesMessageHandler.client2server.registerForMessage(this, MalmoMessageType.CLIENT_INVENTORY_CHANGE);
-    }
-
-    @Override
-    public void deinstall(MissionInit missionInit) {
-        SidesMessageHandler.client2server.deregisterForMessage(this, MalmoMessageType.CLIENT_INVENTORY_CHANGE);
     }
 
 }
